+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fixed unexpected zoom reset when hovering perf rank details (#290)
 -   Fix varying widthes for tables (#298)
 -   Do not display parent task section in drawers if there are no parent tasks (#304)
+-   Fixed font size in popover menus (#300)
 
 ### Changed